glnx-macros.h: add GLNX_HASH_TABLE_FOREACH
It annoyed me how much work it was to simply iterate over a GHashTable,
not to mention polluting the scope. GLNX_HASH_TABLE_FOREACH wraps all
the prep work needed for it.

I realize the implementation itself is not pretty to look at, though the
resulting code is *much* nicer to look at and work with.