fdio: Merge systemd code to use copy_file_range(), drop reflink support
This is the new standard API for this; all Linux filesystems will make use of
this if possible. For example, both NFS and XFS (with reflink enabled)
understand it.

Part of the reason I'm doing this is so that ostree's `/etc` merge will start
using XFS reflinks. But another major reason is to take the next step after and
copy this code into GLib as well, so that all of the general GLib users will
benefit; e.g. Nautilus will transparently do server copy offloads with NFS home
directories.

At the same time, I don't see a reason to still use reflinks. Systemd's code has
an option for this, and everything seems to use `COPY_REFLINK`; so not sure why
they exposed the option.  AFAICS, since upstream Linux
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=a76b5b04375f974579c83433b06466758c0c552c
the btrfs ioctl is really the same thing as `copy_file_range`, so let's just use that.