fdio: Add glnx_fstatat_or_noent()
This is a very common pattern in both ostree/rpm-ostree. Make a better API for
this. I thought a lot about simply zeroing out `struct stat` but that feels
dangerous; none of the values have seem obviously `cannot be zero`.

Rust and Haskell of course have us beat here with the `Option/Maybe` types.